Added all IPA vowels
Added all the IPA vowels per Wikipedia:
https://en.wikipedia.org/wiki/International_Phonetic_Alphabet_chart

Some vowels like /ə/ (schwa) that don't have a "roundedness" dimension listed are marked as "unrounded." Ask someone if you should add a "N/A" category for roundedness that just applied to schwa and /ɐ/.

The GUI window closes and reopens for every letter refresh. Is this the way to do it, or should the window just update? Find out how to do this.